move functional test
